feat: refactor support help components with improved navigation and loading handling
