+++ web/components/support-center/help-center.tsx
@@ -23,12 +23,14 @@ import { MdPreview } from "md-editor-rt"
 import Link from "next/link"
 import { useParams, usePathname, useRouter, useSearchParams } from "next/navigation"
 import { useTheme } from "next-themes"
-import { useCallback, useEffect, useMemo, useRef, useState, type MouseEvent as ReactMouseEvent, type ReactNode } from "react"
+import { useCallback, useEffect, useMemo, useRef, useState, type ReactNode } from "react"
 import { toast } from "sonner"
 
 import { OptionCombobox } from "@/components/option-combobox"
 import { useImageLightboxOptional } from "@/components/image-lightbox"
 import { SafeRichHTML } from "@/components/safe-rich-html"
+import { SupportHelpLink, type HelpPageNavigationHandler, useSupportHelpRoute } from "@/components/support-center/support-help-navigation"
+import { useSupportHelpReader } from "@/components/support-center/use-support-help-reader"
 import { Badge } from "@/components/ui/badge"
 import { Button, buttonVariants } from "@/components/ui/button"
 import { Card, CardContent } from "@/components/ui/card"
@@ -39,7 +41,6 @@ import {
   acceptSupportAnswer,
   createSupportAnswer,
   createSupportQuestion,
-  fetchSupportHelpPage,
   fetchSupportHelpPages,
   fetchSupportMe,
   fetchSupportQuestion,
@@ -60,18 +61,6 @@ import { cn, formatDateTime } from "@/lib/utils"
 
 const pageBackground =
   "bg-[#f7f9fc] dark:bg-background"
-
-type HelpPageNavigationHandler = (event: ReactMouseEvent<HTMLElement>, page: SupportHelpPage) => void
-
-function supportHelpPageHref(slugOrId: string | number) {
-  return `/support/help/${encodeURIComponent(String(slugOrId))}/`
-}
-
-function helpSlugFromPathname(pathname: string) {
-  const parts = pathname.split("/").filter(Boolean)
-  const helpIndex = parts.findIndex((part, index) => part === "help" && parts[index - 1] === "support")
-  return helpIndex >= 0 ? decodeURIComponent(parts[helpIndex + 1] || "") : ""
-}
 
 export function SupportHelpCenter() {
   const t = useI18n()
@@ -187,105 +176,16 @@ function SupportHelpReader() {
   const t = useI18n()
   const { resolvedTheme } = useTheme()
   const pathname = usePathname()
-  const [page, setPage] = useState<SupportHelpPage | null>(null)
-  const [pages, setPages] = useState<SupportHelpPage[]>([])
-  const [activeSlug, setActiveSlug] = useState(() => helpSlugFromPathname(pathname))
   const [query, setQuery] = useState("")
   const [searchResults, setSearchResults] = useState<SupportHelpPage[]>([])
   const [searchLoading, setSearchLoading] = useState(false)
-  const [navigationLoading, setNavigationLoading] = useState(true)
-  const [pageLoading, setPageLoading] = useState(true)
-  const [failed, setFailed] = useState(false)
   const [navigationOpen, setNavigationOpen] = useState(false)
-  const [expanded, setExpanded] = useState<Set<number>>(new Set())
-  const initialSlug = useRef(activeSlug)
-  const detailCache = useRef(new Map<string, SupportHelpPage>())
-  const detailRequestId = useRef(0)
-  const expandedInitialized = useRef(false)
-
-  useEffect(() => {
-    void fetchSupportHelpPages({ limit: 500 })
-      .then((list) => {
-        setFailed(false)
-        setPages(list.results)
-        const target = initialSlug.current || list.results[0]?.slug || String(list.results[0]?.id || "")
-        if (!target) {
-          setPage(null)
-          setExpanded(new Set())
-          setPageLoading(false)
-          return
-        }
-        if (!initialSlug.current) {
-          window.history.replaceState(null, "", supportHelpPageHref(target))
-          setActiveSlug(target)
-        }
-      })
-      .catch(() => {
-        setPage(null)
-        setFailed(true)
-        setPageLoading(false)
-      })
-      .finally(() => setNavigationLoading(false))
-  }, [])
-
-  useEffect(() => {
-    if (!activeSlug || !pages.length) return
-    const requestId = ++detailRequestId.current
-    const cached = detailCache.current.get(activeSlug)
-    const detailRequest = cached ? Promise.resolve(cached) : fetchSupportHelpPage(activeSlug)
-    void detailRequest
-      .then((detail) => {
-        if (requestId !== detailRequestId.current) return
-        detailCache.current.set(activeSlug, detail)
-        if (detail.slug) detailCache.current.set(detail.slug, detail)
-        detailCache.current.set(String(detail.id), detail)
-        setFailed(false)
-        setPage(detail)
-        const activePath = [...helpPageAncestorIds(pages, detail.id), detail.id]
-        setExpanded((current) => {
-          if (!expandedInitialized.current) {
-            expandedInitialized.current = true
-            return new Set(activePath)
-          }
-          const next = new Set(current)
-          activePath.forEach((id) => next.add(id))
-          return next
-        })
-      })
-      .catch(() => {
-        if (requestId !== detailRequestId.current) return
-        setFailed(true)
-      })
-      .finally(() => {
-        if (requestId === detailRequestId.current) setPageLoading(false)
-      })
-  }, [activeSlug, pages])
-
-  useEffect(() => {
-    const handlePopState = () => {
-      setPageLoading(true)
-      setActiveSlug(helpSlugFromPathname(window.location.pathname))
-    }
-    window.addEventListener("popstate", handlePopState)
-    return () => window.removeEventListener("popstate", handlePopState)
-  }, [])
-
+  const { activeSlug, navigate, replace } = useSupportHelpRoute(pathname)
+  const { page, pages, expanded, setExpanded, navigationLoading, pageLoading, failed } = useSupportHelpReader(activeSlug, replace)
   const navigateToHelpPage = useCallback<HelpPageNavigationHandler>((event, targetPage) => {
-    if (event.defaultPrevented || event.button !== 0 || event.metaKey || event.ctrlKey || event.shiftKey || event.altKey) return
-    event.preventDefault()
-    const target = targetPage.slug || String(targetPage.id)
-    const href = supportHelpPageHref(target)
-    if (window.location.pathname === href) {
-      setNavigationOpen(false)
-      window.scrollTo({ top: 0, behavior: "auto" })
-      return
-    }
-    window.history.pushState(null, "", href)
-    setPageLoading(true)
-    setActiveSlug(target)
-    setNavigationOpen(false)
-    window.scrollTo({ top: 0, behavior: "auto" })
-  }, [])
+    navigate(event, targetPage)
+    if (event.defaultPrevented) setNavigationOpen(false)
+  }, [navigate])
 
   useEffect(() => {
     if (!page) return
@@ -325,7 +225,7 @@ function SupportHelpReader() {
     <SupportDocsFrame
       navigationOpen={navigationOpen}
       onNavigationOpenChange={setNavigationOpen}
-      navigation={<HelpNavigation pages={visiblePages} rootPages={visiblePages.filter((item) => !item.parentId)} searchResults={searchResults} title={query} expanded={expanded} selectedPageId={page?.id ?? 0} loading={navigationLoading || searchLoading} failed={failed} onTitleChange={(value) => { setQuery(value); setSearchResults([]); setSearchLoading(Boolean(value.trim())) }} onExpandedChange={setExpanded} onNavigate={navigateToHelpPage} linkMode />}
+      navigation={<HelpNavigation pages={visiblePages} rootPages={visiblePages.filter((item) => !item.parentId)} searchResults={searchResults} title={query} expanded={expanded} selectedPageId={page?.id ?? 0} loading={navigationLoading || searchLoading} failed={failed} onTitleChange={(value) => { setQuery(value); setSearchResults([]); setSearchLoading(Boolean(value.trim())) }} onExpandedChange={setExpanded} onNavigate={navigateToHelpPage} />}
       toc={<PublicArticleToc content={page?.content ?? ""} contentType={page?.contentType} />}
     >
       <div aria-busy={pageLoading} className={cn(page && pageLoading && "opacity-60 transition-opacity")}>
@@ -699,7 +599,6 @@ function HelpNavigation({
   onTitleChange,
   onExpandedChange,
   onNavigate,
-  linkMode = false,
 }: {
   pages: SupportHelpPage[]
   rootPages: SupportHelpPage[]
@@ -712,7 +611,6 @@ function HelpNavigation({
   onTitleChange: (value: string) => void
   onExpandedChange: (value: Set<number>) => void
   onNavigate: HelpPageNavigationHandler
-  linkMode?: boolean
 }) {
   const t = useI18n()
   return (
@@ -720,17 +618,17 @@ function HelpNavigation({
       <SupportSearchInput value={title} onChange={onTitleChange} placeholder={t("supportPublic.help.searchPlaceholder")} compact />
       <div className="mt-4 grid gap-0.5">
         {title.trim() ? searchResults.map((page) => (
-          <a key={page.id} href={supportHelpPageHref(page.slug || page.id)} onClick={(event) => onNavigate(event, page)} className={cn("rounded-lg px-2.5 py-2 text-sm transition-colors hover:bg-muted", selectedPageId === page.id && "bg-primary/10 text-primary")}>
+          <SupportHelpLink key={page.id} page={page} onNavigate={onNavigate} className={cn("rounded-lg px-2.5 py-2 text-sm transition-colors hover:bg-muted", selectedPageId === page.id && "bg-primary/10 text-primary")}>
             <span className="block truncate font-medium">{page.title}</span>
             {page.summary ? <span className="mt-1 block line-clamp-2 text-xs leading-5 text-muted-foreground">{page.summary}</span> : null}
-          </a>
+          </SupportHelpLink>
         )) : rootPages.map((page) => (
           <PublicHelpPageNode key={page.id} page={page} depth={0} pages={pages} expanded={expanded} selectedPageId={selectedPageId} onToggle={(id) => {
             const next = new Set(expanded)
             if (next.has(id)) next.delete(id)
             else next.add(id)
             onExpandedChange(next)
-          }} onNavigate={onNavigate} linkMode={linkMode} />
+          }} onNavigate={onNavigate} />
         ))}
         {loading ? <div className="px-2 py-8 text-center text-sm text-muted-foreground">{t("supportPublic.loading.navigation")}</div> : null}
         {!loading && (title.trim() ? !searchResults.length : !pages.length) ? <EmptyState text={failed ? t("supportPublic.empty.pagesFailed") : t("supportPublic.empty.noPagesMatched")} compact /> : null}
@@ -818,10 +716,10 @@ function HelpArticle({ page, pages, previewId, theme, onNavigate }: { page: Supp
 
 function ArticlePager({ page, direction, onNavigate }: { page: SupportHelpPage; direction: "previous" | "next"; onNavigate: HelpPageNavigationHandler }) {
   const t = useI18n()
-  return <a href={supportHelpPageHref(page.slug || page.id)} onClick={(event) => onNavigate(event, page)} className={cn("group rounded-xl border px-4 py-3 transition-colors hover:border-primary/40 hover:bg-muted/50", direction === "next" && "text-right")}>
+  return <SupportHelpLink page={page} onNavigate={onNavigate} className={cn("group rounded-xl border px-4 py-3 transition-colors hover:border-primary/40 hover:bg-muted/50", direction === "next" && "text-right")}>
     <span className="text-xs text-muted-foreground">{t(`supportPublic.help.${direction}`)}</span>
     <span className="mt-1 flex items-center justify-between gap-3 text-sm font-medium text-primary">{direction === "previous" ? <ChevronRightIcon className="size-4 rotate-180" /> : null}<span className={cn("truncate", direction === "next" && "ml-auto")}>{page.title}</span>{direction === "next" ? <ChevronRightIcon className="size-4" /> : null}</span>
-  </a>
+  </SupportHelpLink>
 }
 
 function SupportEntryCard({
@@ -916,7 +814,6 @@ function PublicHelpPageNode({
   selectedPageId,
   onToggle,
   onNavigate,
-  linkMode = false,
 }: {
   page: SupportHelpPage
   depth: number
@@ -925,7 +822,6 @@ function PublicHelpPageNode({
   selectedPageId: number
   onToggle: (id: number) => void
   onNavigate: HelpPageNavigationHandler
-  linkMode?: boolean
 }) {
   const t = useI18n()
   const open = expanded.has(page.id)
@@ -953,42 +849,22 @@ function PublicHelpPageNode({
             {open ? <ChevronDownIcon className="size-4" /> : <ChevronRightIcon className="size-4" />}
           </button>
         ) : <span className="size-7 shrink-0" />}
-        {linkMode ? (
-          <a href={supportHelpPageHref(page.slug || page.id)} className="flex min-w-0 flex-1 items-center gap-2 py-1.5 text-left" onClick={(event) => onNavigate(event, page)} aria-current={selected ? "page" : undefined}>
-            {hasChildren ? (open ? <FolderOpenIcon className="size-4 shrink-0" /> : <FolderIcon className="size-4 shrink-0" />) : <FileTextIcon className="size-3.5 shrink-0 opacity-70" />}
-            <span className="truncate">{page.title}</span>
-          </a>
-        ) : (
-          <button type="button" className="flex min-w-0 flex-1 items-center gap-2 py-1.5 text-left" onClick={(event) => onNavigate(event, page)} aria-current={selected ? "page" : undefined}>
-            {hasChildren ? (open ? <FolderOpenIcon className="size-4 shrink-0" /> : <FolderIcon className="size-4 shrink-0" />) : <FileTextIcon className="size-3.5 shrink-0 opacity-70" />}
-            <span className="truncate">{page.title}</span>
-          </button>
-        )}
+        <SupportHelpLink page={page} onNavigate={onNavigate} className="flex min-w-0 flex-1 items-center gap-2 py-1.5 text-left" aria-current={selected ? "page" : undefined}>
+          {hasChildren ? (open ? <FolderOpenIcon className="size-4 shrink-0" /> : <FolderIcon className="size-4 shrink-0" />) : <FileTextIcon className="size-3.5 shrink-0 opacity-70" />}
+          <span className="truncate">{page.title}</span>
+        </SupportHelpLink>
       </div>
       {open && hasChildren ? (
         <div className="relative before:absolute before:inset-y-1 before:w-px before:bg-border/80" style={{ marginLeft: `${depth * 16 + 17}px` }}>
           <div style={{ marginLeft: `${-(depth * 16 + 17)}px` }}>
             {children.map((child) => (
-              <PublicHelpPageNode key={child.id} page={child} depth={depth + 1} pages={pages} expanded={expanded} selectedPageId={selectedPageId} onToggle={onToggle} onNavigate={onNavigate} linkMode={linkMode} />
+              <PublicHelpPageNode key={child.id} page={child} depth={depth + 1} pages={pages} expanded={expanded} selectedPageId={selectedPageId} onToggle={onToggle} onNavigate={onNavigate} />
             ))}
           </div>
         </div>
       ) : null}
     </div>
   )
-}
-
-function helpPageAncestorIds(pages: SupportHelpPage[], pageId: number) {
-  const ancestors: number[] = []
-  const pagesById = new Map(pages.map((item) => [item.id, item]))
-  const visited = new Set<number>()
-  let parentId = pagesById.get(pageId)?.parentId ?? 0
-  while (parentId && !visited.has(parentId)) {
-    ancestors.push(parentId)
-    visited.add(parentId)
-    parentId = pagesById.get(parentId)?.parentId ?? 0
-  }
-  return ancestors
 }
 
 function ChildPageLinks({ pages, onNavigate }: { pages: SupportHelpPage[]; onNavigate: HelpPageNavigationHandler }) {
@@ -999,10 +875,10 @@ function ChildPageLinks({ pages, onNavigate }: { pages: SupportHelpPage[]; onNav
       <h3 className="mb-3 font-semibold">{t("supportPublic.help.childPages")}</h3>
       <div className="grid gap-2 sm:grid-cols-2">
         {pages.map((page) => (
-          <a key={page.id} href={supportHelpPageHref(page.slug || page.id)} onClick={(event) => onNavigate(event, page)} className="rounded-xl border p-3 transition hover:bg-muted/60">
+          <SupportHelpLink key={page.id} page={page} onNavigate={onNavigate} className="rounded-xl border p-3 transition hover:bg-muted/60">
             <span className="font-medium">{page.title}</span>
             {page.summary ? <span className="mt-1 block text-sm text-muted-foreground">{page.summary}</span> : null}
-          </a>
+          </SupportHelpLink>
         ))}
       </div>
     </div>
